fix: reduced files in pages_utils
